+++ src/App.js
@@ -16,7 +16,7 @@ import LectureUploadPage from './pages/Lecture/LectureUploadPage';
 import MentoringPage from './pages/Mentoring/MentoringPage';
 import MentoringUploadPage from './pages/Mentoring/MentoringUploadPage';
 import MentoringDetailPage from './pages/Mentoring/MentoringDetailPage';
-import MentoringChat from './components/MentoringPage/MentoringChat';
+import MentoringChatPage from './pages/Mentoring/MentoringChatPage';
 
 
 
@@ -68,8 +68,8 @@ function Mentoring() {
       <Routes>
           <Route path='' element={<MentoringPage/>}/>
           <Route path='upload' element={<MentoringUploadPage/>}/>
-          <Route path='detail' element={<MentoringDetailPage/>}/>
-          <Route path='detail/:id/chating/:id' element={<MentoringChat/>}/>
+          <Route path='detail/:id' element={<MentoringDetailPage/>}/>
+          <Route path='detail/:id/chating/:id' element={<MentoringChatPage/>}/>
       </Routes>
   );
 };
